+++ pybamm/models/submodels/external_circuit/base_external_circuit.py
@@ -12,51 +12,36 @@ class BaseModel(pybamm.BaseSubModel):
 
     def get_fundamental_variables(self):
         Q_Ah = pybamm.standard_variables.Q_Ah
-
-        variables = {"Discharge capacity [A.h]": Q_Ah}
-        if self.options["calculate discharge energy"] == "true":
-            Q_Wh = pybamm.standard_variables.Q_Wh
-            Qt_Wh = pybamm.standard_variables.Qt_Wh
-            Qt_Ah = pybamm.standard_variables.Qt_Ah
-            variables.update(
-                {
-                    "Discharge energy [W.h]": Q_Wh,
-                    "Throughput energy [W.h]": Qt_Wh,
-                    "Throughput capacity [A.h]": Qt_Ah,
-                }
-            )
-        else:
-            variables.update(
-                {
-                    "Discharge energy [W.h]": pybamm.Scalar(0),
-                    "Throughput energy [W.h]": pybamm.Scalar(0),
-                    "Throughput capacity [A.h]": pybamm.Scalar(0),
-                }
-            )
+        Q_Wh = pybamm.standard_variables.Q_Wh
+        Qt_Wh = pybamm.standard_variables.Qt_Wh
+        Qt_Ah = pybamm.standard_variables.Qt_Ah
+        variables = {
+            "Discharge capacity [A.h]": Q_Ah,
+            "Discharge energy [W.h]": Q_Wh,
+            "Throughput energy [W.h]": Qt_Wh,
+            "Throughput capacity [A.h]": Qt_Ah,
+        }
         return variables
 
     def set_initial_conditions(self, variables):
-        Q_Ah = variables["Discharge capacity [A.h]"]
-        self.initial_conditions[Q_Ah] = pybamm.Scalar(0)
-        if self.options["calculate discharge energy"] == "true":
-            Q_Wh = variables["Discharge energy [W.h]"]
-            Qt_Wh = variables["Throughput energy [W.h]"]
-            Qt_Ah = variables["Throughput capacity [A.h]"]
-            self.initial_conditions[Q_Wh] = pybamm.Scalar(0)
-            self.initial_conditions[Qt_Wh] = pybamm.Scalar(0)
-            self.initial_conditions[Qt_Ah] = pybamm.Scalar(0)
+        for name in [
+            "Discharge capacity [A.h]",
+            "Discharge energy [W.h]",
+            "Throughput energy [W.h]",
+            "Throughput capacity [A.h]",
+        ]:
+            var = variables[name]
+            self.initial_conditions[var] = pybamm.Scalar(0)
 
     def set_rhs(self, variables):
-        # ODEs for discharge capacity and throughput capacity
-        Q_Ah = variables["Discharge capacity [A.h]"]
+        # ODEs for discharge/throughput capacity/energy
         I = variables["Current [A]"]
-
-        self.rhs[Q_Ah] = I * self.param.timescale / 3600
-        if self.options["calculate discharge energy"] == "true":
-            Q_Wh = variables["Discharge energy [W.h]"]
-            Qt_Wh = variables["Throughput energy [W.h]"]
-            Qt_Ah = variables["Throughput capacity [A.h]"]
-            V = variables["Terminal voltage [V]"]
-            self.rhs[Q_Wh] = I * V * self.param.timescale / 3600
-            self.rhs[Qt_Wh] = abs(I * V) * self.param.timescale / 3600
-            self.rhs[Qt_Ah] = abs(I) * self.param.timescale / 3600
+        V = variables["Terminal voltage [V]"]
+        for name, integrand in [
+            ("Discharge capacity [A.h]", I),
+            ("Throughput capacity [A.h]", abs(I)),
+            ("Discharge energy [W.h]", I * V),
+            ("Throughput energy [W.h]", abs(I * V)),
+        ]:
+            var = variables[name]
+            self.rhs[var] = integrand * self.param.timescale / 3600