+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fix installation on `Google Colab` (`pybtex` and `Colab` issue) ([#2526](https://github.com/pybamm-team/PyBaMM/pull/2526))
 
+## Breaking changes
+
+- Option "calculate discharge energy" has been removed. Discharge energy is now always calculated
+
 # [v22.11](https://github.com/pybamm-team/PyBaMM/tree/v22.11) - 2022-11-30
 
 ## Features